backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-genai
python-dotenv
gTTS